spinner18

+++ app/src/main/java/com/step/envocab/CountDao.java
@@ -17,19 +17,19 @@ public interface CountDao {
     List<Dbcounts> getCounts(int id_exercise);
 
     @Query("SELECT id_group from Dbcounts WHERE id_exercice=:id_exercise ORDER BY trainDate DESC  ")
-    long lastGroup(int id_exercise);
+    int lastGroup(int id_exercise);
 
     @Query("SELECT * from Dbcounts WHERE id_exercice = :id_exercise AND id_word=:id_word AND id_group=:id_group ORDER BY id_count")
-    Dbcounts getCountsFiltered(Integer id_exercise, Integer id_word, Long id_group);
+    Dbcounts getCountsFiltered(Integer id_exercise, Integer id_word, int id_group);
 
     //@Insert(onConflict = OnConflictStrategy.REPLACE)
     @Query("insert into Dbcounts ('id_exercice','id_word','id_group','train','trainDate') VALUES(:id_exercise,:id_word,:id_group,:train,:train_date)")
-    Long insCount( Integer id_exercise, Integer id_word, Long id_group, Boolean train, Date train_date);
+    Long insCount( Integer id_exercise, Integer id_word, int id_group, Boolean train, Date train_date);
 
     @Query("update Dbcounts set id_exercice = :id_exercise, id_word=:id_word,  id_group =:id_group, train =:train, trainDate=:train_date WHERE id_count = :id_count")
-    int upCount(Integer id_count, Integer id_exercise, Integer id_word, Long id_group, Boolean train, Date train_date);
+    int upCount(Integer id_count, Integer id_exercise, Integer id_word, int id_group, Boolean train, Date train_date);
 
-    default void insertOrUpdate(Integer id_exercise, Integer id_word, Long id_group, Boolean train, Date train_date) {
+    default void insertOrUpdate(Integer id_exercise, Integer id_word, int id_group, Boolean train, Date train_date) {
         Dbcounts countsFromDB = getCountsFiltered(id_exercise, id_word,  id_group);
         if (countsFromDB!=null) {
             upCount(countsFromDB.getId_count(),id_exercise, id_word,  id_group, train, train_date );
